requirements.txt
git+https://github.com/facebookresearch/segment-anything.git
opencv-python
pycocotools 
matplotlib 
onnxruntime 
onnx
tqdm
# torch
# torchvision
pyYAML
pandas
#torch==2.0.0+cu117 torchvision==0.15.0+cu117 -f https://download.pytorch.org/whl/torch_stable.html
